requirements.txt
matplotlib==3.4.1
pandas==1.2.1
tensorflow==2.4.1
regex==2021.3.17
numpy==1.19.4
arrow==1.0.3
beautifulsoup4==4.9.3
python_dateutil==2.8.1
